+++ ui-smartlib-web/src/app/[locale]/checkout/page.tsx
@@ -1,192 +1,106 @@
 "use client";
 
-import { Button } from "@/components/ui/button";
-import { Link, useRouter } from "@/config/i18n/routing";
-import useGetBooks from "@/hooks/useGetBooks";
-import { removeItem, updateItem } from "@/redux/features/cartSlice";
-import { useAppDispatch, useAppSelector } from "@/redux/hooks";
-import { ArrowLeft } from "lucide-react";
-import { useLocale, useTranslations } from "next-intl";
+import { redirect } from "@/config/i18n/routing";
+import { useAppSelector } from "@/redux/hooks";
+import { useTranslations } from "next-intl";
 import Image from "next/image";
-import { useEffect } from "react";
+import { useEffect, useMemo, useState } from "react";
 
-const CartItem = ({ item, buttonTextTl, onButtonClick, onTrashClick }) => {
-  const locale = useLocale();
-  const t = useTranslations("reviewMyCart");
+export default function CheckoutPage() {
+  const cartItems = useAppSelector((state) => state.cart.items);
+  const [currentIndex, setCurrentIndex] = useState(0);
+  const t = useTranslations("checkout");
 
-  return (
-    <article key={item?.isbn} className="relative flex justify-between gap-3">
-      <div className="flex gap-5 items-center">
-        <figure>
-          <Image
-            width={300}
-            height={200}
-            src={item.image}
-            alt="book item in cart"
-            className="w-[110px] h-[135px] rounded-md object-cover"
-          />
-        </figure>
-        <div className="space-y-1">
-          <p className="text-xs bg-primary px-2 py-1 rounded-sm font-medium text-white w-fit">
-            {item.category?.[`${locale}Category`]}
-          </p>
-          <h4 className="text-2xl font-medium">{item.title}</h4>
-          <p className="text-secondary-foreground">{item.author}</p>
-        </div>
-      </div>
-      <div className="flex items-center gap-4">
-        <Button
-          onClick={() => onButtonClick(item)}
-          className="bg-accent text-primary rounded-full px-6 py-6 shadow-sm"
-        >
-          {t(buttonTextTl)}
-        </Button>
-        <button onClick={() => onTrashClick(item)}>
-          <Image
-            width={40}
-            height={40}
-            src="/icons/trash.svg"
-            alt="Remove book from cart"
-            className="w-[30px] h-[30px] rounded-md object-cover"
-          />
-        </button>
-      </div>
-    </article>
-  );
-};
+  useEffect(() => {
+    // if (!cartItems?.length) redirect("/home");
+    //mutation every time current index changes, to open book door.
+  }, [cartItems]);
 
-export default function ReviewCartPage() {
-  const { rentItems, purchaseItems } = useGetBooks();
-  const settings = useAppSelector((state) => state.settings);
-  const t = useTranslations("cart");
-  const tReviewCart = useTranslations("reviewMyCart");
-
-  const router=useRouter()
-
-  const dispatch = useAppDispatch();
-
-  const handleMoveToBuy = (isbn: string) => {
-    dispatch(updateItem({ isbn, transactionType: "PURCHASE" }));
-  };
-
-  const handleMoveToRent = (isbn: string) => {
-    dispatch(updateItem({ isbn, transactionType: "RENT" }));
-  };
-
-  const handleRemoveItem = (isbn: string) => {
-    dispatch(removeItem({ isbn }));
-  };
-
-  useEffect(()=>{
-    if(rentItems.length || purchaseItems.length)return
-    router.push('/catalog')
-  },[rentItems,purchaseItems,router])
+  const currentBook = useMemo(() => {
+    return cartItems[currentIndex];
+  }, [currentIndex, cartItems]);
 
   return (
-    <section className="grid grid-rows-[auto_1fr,auto] h-screen">
-      <div className="flex justify-between items-center px-6 py-3 bg-white">
-        <Link href={`/catalog`}>
-          <div className="flex gap-5 items-center">
-            <ArrowLeft />
-            <p className=" first-letter:uppercase font-medium">
-              {tReviewCart("myCart")}
+    <div>
+      <header className="bg-white py-5 flex justify-center">
+        <p className="text-base font-medium ml-[190px]">
+          {t("readInstructions")}
+        </p>
+      </header>
+      <div className="grid grid-cols-[auto_1fr]">
+        <aside className="bg-white space-y-6 w-[200px] pt-5 px-3">
+          <div>
+            <h3 className="text-2xl font-semibold">{t("yourBooks")}</h3>
+          </div>
+          <div className="space-y-4">
+            {cartItems.map((item, index) => (
+              <article key={item.isbn}>
+                <p className="font-medium">
+                  {index + 1}. {item.title}
+                </p>
+              </article>
+            ))}
+          </div>
+        </aside>
+        <section className="grid grid-rows-[1fr_auto place-content-center gap-10 mt-10  text-center">
+          <div className="space-y-5">
+            <div>
+              <h2 className="text-3xl font-medium">{t("pickYourBook")}</h2>
+            </div>
+            <div>
+              <figure>
+                <Image
+                  src={currentBook?.image}
+                  width={250}
+                  height={350}
+                  alt={`${currentBook?.title} image`}
+                  className="m-auto rounded-md w-[210px] h-[270px]"
+                />
+              </figure>
+            </div>
+            <div>
+              <h3 className="text-4xl font-medium">
+                {t("openingDoor1")}{" "}
+                <span className="text-primary ">{t("openingDoor2")}</span>{" "}
+                <span className="text-primary">(3B)</span>
+              </h3>
+            </div>
+            <div className="text-left flex gap-3 bg-white p-5 rounded-2xl max-w-[500px]">
+              <Image
+                src="/images/door-checkout.svg"
+                width={250}
+                height={350}
+                className="w-[160px] h-[170px] object-top object-cover runded-md"
+                alt="door checkout image"
+              />
+              <div className="space-y-5">
+                <div className="space-y-2">
+                  <h4 className="font-medium">{t("firstStepTitle")}</h4>
+                  <p>
+                    {t("firstStepContent1")}
+                    <span className="text-primary">
+                      {" "}
+                      {t("firstStepContent2")}
+                      3B
+                    </span>{" "}
+                    {t("firstStepContent3")}
+                  </p>
+                </div>
+                <div className="space-y-2">
+                  <h4 className="font-medium"> {t("nextStepTitle")}</h4>
+                  <p> {t("nextStepContent")}</p>
+                </div>
+              </div>
+            </div>
+          </div>
+          <div className="pb-2">
+            <p className="font-medium space-x-1">
+              <span> {t("closeDoorTime1")}</span> <span>1</span><span>{t("closeDoorTime2")}</span> 3
+              <span>{t("closeDoorTime3")}</span>
             </p>
           </div>
-        </Link>
-
-        <div>
-          <figure>
-            <Image
-              width={150}
-              height={70}
-              src="/images/releevante.svg"
-              alt="Remove book from cart"
-              className="w-[120px] h-auto rounded-md object-cover"
-            />
-          </figure>
-        </div>
+        </section>
       </div>
-      <div className="overflow-y-auto px-4 py-4 space-y-6">
-        {rentItems?.length > 0 && (
-          <div className="pt-7 grid bg-white rounded-xl space-y-5">
-            <div className="px-4">
-              <h3 className="space-x-1 font-medium flex">
-                <p className="space-x-1">
-                  <span className="capitalize">{t("to")}</span>
-                  <span className="text-black">{t("rent")} </span>
-                </p>
-                <p className="space-x-1 text-secondary-foreground">
-                  (<span>{rentItems.length}</span>
-                  {settings?.data && (
-                    <>
-                      <span>{t("of")}</span>
-                      <span>{settings?.data?.maxBooksPerLoan}</span>
-                    </>
-                  )}
-                  )
-                </p>
-              </h3>
-            </div>
-            <div className="px-4 space-y-4">
-              {rentItems.map((item) => (
-                <CartItem
-                  key={item.isbn}
-                  item={item}
-                  buttonTextTl="moveToBuy"
-                  onButtonClick={() => handleMoveToBuy(item.isbn)}
-                  onTrashClick={() => handleRemoveItem(item.isbn)}
-                />
-              ))}
-            </div>
-            <div className="flex justify-center items-center border-t border-secondary  py-3 px-5 bg-white">
-              <Link
-                href={'/catalog'}
-                className="m-auto border rounded-full font-medium tracking-wider text-sm py-4 px-7 border-primary text-primary bg-transparent"
-              >
-                {tReviewCart("rentAnotherBook")}
-              </Link>
-            </div>
-          </div>
-        )}
-        {purchaseItems?.length > 0 && (
-          <div className="pt-7 pb-5 px-5 bg-white rounded-xl space-y-5">
-            <div>
-              <h3 className="space-x-1 font-medium flex text-secondary-foreground">
-                <p className="space-x-1">
-                  <span className="capitalize text-black">{t("to")}</span>
-                  <span className="text-black">{t("purchase")}</span>
-                </p>
-                <p className="space-x-1">
-                  (<span>{purchaseItems?.length}</span>
-                  {settings?.data && (
-                    <>
-                      <span>{t("of")}</span>
-                      <span>{settings?.data?.maxBooksPerLoan}</span>
-                    </>
-                  )}
-                  )
-                </p>
-              </h3>
-            </div>
-            <div className="space-y-4">
-              {purchaseItems.map((item) => (
-                <CartItem
-                  key={item.isbn}
-                  item={item}
-                  buttonTextTl="moveToRent"
-                  onButtonClick={() => handleMoveToRent(item.isbn)}
-                  onTrashClick={() => handleRemoveItem(item.isbn)}
-                />
-              ))}
-            </div>
-          </div>
-        )}
-      </div>
-      <div className="flex justify-center items-center  py-3 px-5 bg-white">
-        <Button className="m-auto rounded-full py-6 px-7 hover:text-white border-primary">
-          <span className="first-letter:uppercase"> {tReviewCart("confirmWithdrawal")}</span>
-        </Button>
-      </div>
-    </section>
+    </div>
   );
 }
